backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
yfinance==0.2.43
financedatareader==0.9.50
pandas==2.2.2
pandas-ta==0.3.14b0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
slowapi==0.1.9
PyJWT==2.9.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic==2.9.2
